app/python/requirements.txt
tk
customtkinter
pyserial
customtkinter
